bot/requirements.txt
apache-libcloud==3.8.0
black==24.1.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
fasteners==0.19
greenlet==3.0.3
idna==3.6
isort==5.13.2
loguru==0.7.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
psycopg2==2.9.9
pyTelegramBotAPI==4.15.3
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.25
sqlalchemy-file==0.6.0
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
